django-pictures-with-filters/requirements.txt
Django==4.0.4
opencv-python==4.5.5.64
Pillow==9.1.1
django-crispy-forms==1.14.0